slack_agent/requirements.txt
slack_sdk[socket_mode]
python-dotenv
requests
